Add AnnotateModels::SchemaInfo::ColInfo
