+++ lib/generators/annotate/templates/auto_annotate_models.rake
@@ -34,8 +34,8 @@ if Rails.env.development?
       'ignore_columns'            => nil,
       'ignore_routes'             => nil,
       'ignore_unknown_models'     => 'false',
-      'hide_limit_column_types'   => '<%= AnnotateModels::NO_LIMIT_COL_TYPES.join(",") %>',
-      'hide_default_column_types' => '<%= AnnotateModels::NO_DEFAULT_COL_TYPES.join(",") %>',
+      'hide_limit_column_types'   => '<%= AnnotateModels::SchemaInfo::ColInfo::NO_LIMIT_COL_TYPES.join(",") %>',
+      'hide_default_column_types' => '<%= AnnotateModels::SchemaInfo::ColInfo::NO_DEFAULT_COL_TYPES.join(",") %>',
       'skip_on_db_migrate'        => 'false',
       'format_bare'               => 'true',
       'format_rdoc'               => 'false',
