+++ spec/annotate/annotate_models_spec.rb
@@ -1759,17 +1759,17 @@ end
     end
   end
 
-  describe 'SchemaInfo' do
+  describe 'SchemaInfo::ColInfo' do
     describe '.#quote' do
       it 'converts values to the default column value as a string' do
-        expect(AnnotateModels::SchemaInfo.send(:quote, nil)).to eql('NULL')
-        expect(AnnotateModels::SchemaInfo.send(:quote, true)).to eql('TRUE')
-        expect(AnnotateModels::SchemaInfo.send(:quote, false)).to eql('FALSE')
-        expect(AnnotateModels::SchemaInfo.send(:quote, 25)).to eql('25')
-        expect(AnnotateModels::SchemaInfo.send(:quote, 25.6)).to eql('25.6')
-        expect(AnnotateModels::SchemaInfo.send(:quote, 1e-20)).to eql('1.0e-20')
-        expect(AnnotateModels::SchemaInfo.send(:quote, BigDecimal.new('1.2'))).to eql('1.2')
-        expect(AnnotateModels::SchemaInfo.send(:quote, [BigDecimal.new('1.2')])).to eql(['1.2'])
+        expect(AnnotateModels::SchemaInfo::ColInfo.send(:quote, nil)).to eql('NULL')
+        expect(AnnotateModels::SchemaInfo::ColInfo.send(:quote, true)).to eql('TRUE')
+        expect(AnnotateModels::SchemaInfo::ColInfo.send(:quote, false)).to eql('FALSE')
+        expect(AnnotateModels::SchemaInfo::ColInfo.send(:quote, 25)).to eql('25')
+        expect(AnnotateModels::SchemaInfo::ColInfo.send(:quote, 25.6)).to eql('25.6')
+        expect(AnnotateModels::SchemaInfo::ColInfo.send(:quote, 1e-20)).to eql('1.0e-20')
+        expect(AnnotateModels::SchemaInfo::ColInfo.send(:quote, BigDecimal.new('1.2'))).to eql('1.2')
+        expect(AnnotateModels::SchemaInfo::ColInfo.send(:quote, [BigDecimal.new('1.2')])).to eql(['1.2'])
       end
     end
   end
